+++ go/CHANGELOG.md
@@ -1,3 +1,4 @@
+- Windows updater (PR: keybase/client#1770)
 - Fix passphrase confirmation logic (PR: keybase/client#1752)
 - Fix `keybase passphrase recover` SecretUI, LoginSession issues (PR: keybase/client#1750)
 - Allow `keybase sigs revoke` to work with a prefix, and